requirements.txt
imparaai-checkers
torch~=1.11.0
numpy~=1.22.4
matplotlib~=3.5.2
scipy~=1.8.1